refactor(filter): status→errorCode 매핑을 공유 유틸로 추출
HttpExceptionFilter의 status→ErrorCode 매핑과 메시지 추출을
http-status-error-code.util의 단일 소스(HTTP_STATUS_TO_ERROR_CODE,
extractHttpExceptionMessage)로 분리한다. IdempotencyKeyInterceptor가
실패 스냅샷 캐싱 시 같은 매핑을 미러 복제하던 것을 공유 참조로 대체하기 위한
선행 추출 — 멱등성상 최초 실시간 응답 code와 R3 재반환 code가 분기하면 안 된다.
필터 동작(미매핑 status warn + COMMON_INTERNAL_ERROR 폴백)은 보존.

#132
