+++ src/filter/http-exception.filter.ts
@@ -3,11 +3,14 @@ import {
   Catch,
   ExceptionFilter,
   HttpException,
-  HttpStatus,
 } from '@nestjs/common';
 import { Request, Response } from 'express';
 import { ErrorCode } from '../constant/error-code.enum';
 import { AbstractExceptionFilter } from './abstract-exception.filter';
+import {
+  HTTP_STATUS_TO_ERROR_CODE,
+  extractHttpExceptionMessage,
+} from './http-status-error-code.util';
 
 @Catch(HttpException)
 export class HttpExceptionFilter
@@ -26,23 +29,15 @@ export class HttpExceptionFilter
     this.sendFailureResponse(
       res,
       this.mapToErrorCode(exception.getStatus()),
-      this.getMessage(response),
+      extractHttpExceptionMessage(response),
     );
   }
 
+  // 매핑 테이블은 http-status-error-code.util의 단일 소스를 참조한다
+  // (IdempotencyKeyInterceptor와 공유 — 멱등성상 분기 금지). 미매핑 status만
+  // 경고 후 COMMON_INTERNAL_ERROR로 폴백한다.
   private mapToErrorCode(status: number): ErrorCode {
-    const mapping: Record<number, ErrorCode> = {
-      [HttpStatus.BAD_REQUEST]: ErrorCode.COMMON_BAD_REQUEST,
-      [HttpStatus.UNAUTHORIZED]: ErrorCode.COMMON_UNAUTHORIZED,
-      [HttpStatus.NOT_FOUND]: ErrorCode.COMMON_NOT_FOUND,
-      [HttpStatus.NOT_ACCEPTABLE]: ErrorCode.COMMON_NOT_ACCEPTABLE,
-      [HttpStatus.CONFLICT]: ErrorCode.COMMON_CONFLICT,
-      [HttpStatus.TOO_MANY_REQUESTS]: ErrorCode.COMMON_TOO_MANY_REQUESTS,
-      [HttpStatus.INTERNAL_SERVER_ERROR]: ErrorCode.COMMON_INTERNAL_ERROR,
-      [HttpStatus.SERVICE_UNAVAILABLE]: ErrorCode.COMMON_SERVICE_UNAVAILABLE,
-    };
-
-    const errorCode = mapping[status];
+    const errorCode = HTTP_STATUS_TO_ERROR_CODE[status];
     if (!errorCode) {
       this.logger.warn(
         `Unmapped HttpStatus: ${status}, falling back to COMMON_INTERNAL_ERROR`,
@@ -50,19 +45,5 @@ export class HttpExceptionFilter
       return ErrorCode.COMMON_INTERNAL_ERROR;
     }
     return errorCode;
-  }
-
-  private getMessage(response: string | object): string {
-    if (typeof response === 'string') {
-      return response;
-    }
-
-    const message = (response as Record<string, unknown>).message;
-
-    if (Array.isArray(message)) {
-      return message.join(', ');
-    }
-
-    return typeof message === 'string' ? message : '';
   }
 }
